+++ src/pages/AuthPage/Auth/AuthForm/AuthForm.tsx
@@ -1,4 +1,4 @@
-import { ACCESS_TYPE } from '@/constants';
+import { ACCESS_TYPE } from '@/utils/constants';
 
 import AuthContent from '../AuthContent/AuthContent';
 import Input from '@/components/Input/Input';
