+++ src/utils/hooks/useAccessTypeState.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import { ACCESS_TYPE } from "../../constants";
+import { ACCESS_TYPE } from "../constants";
 
 export const useAccessTypeState = () => {
   const [accessType, setAccessType] = useState(ACCESS_TYPE.LANDING);
